Hw8: move kafka container and kafka properties to IntegrationTest
